+++ java/src/main/java/org/lance/ipc/LanceScanner.java
@@ -84,7 +84,7 @@ public class LanceScanner implements org.apache.arrow.dataset.scanner.Scanner {
 
   static native LanceScanner createScanner(
       Dataset dataset,
-      Optional<List<Integer>> fragmentIds,
+      Optional<List<Long>> fragmentIds,
       Optional<List<String>> columns,
       Optional<ByteBuffer> substraitFilter,
       Optional<String> filter,
